# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Safai Market — Anupurna Traders store management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, cancelled]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "400":
          description: Validation error (e.g. insufficient stock)
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get bill details with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "404":
          description: Not found
  /bills/{id}/cancel:
    post:
      operationId: cancelBill
      tags: [bills]
      summary: Cancel a bill (reverses stock movements)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBillInput"
      responses:
        "200":
          description: Cancelled bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"

  # ─── CUSTOMERS ───────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: hasUdhaar
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with udhaar ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer with ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/payments:
    post:
      operationId: receiveCustomerPayment
      tags: [customers]
      summary: Record a payment received from a customer (reduces udhaar)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UdhaarPaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UdhaarLedgerEntry"

  # ─── SUPPLIERS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Bill:
      type: object
      required: [id, billNumber, totalAmount, cashAmount, upiAmount, udhaarAmount, status, createdAt]
      properties:
        id:
          type: integer
        billNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        cashAmount:
          type: number
        upiAmount:
          type: number
        udhaarAmount:
          type: number
        discountAmount:
          type: number
        status:
          type: string
          enum: [active, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        itemCount:
          type: integer
    BillDetail:
      type: object
      required: [id, billNumber, totalAmount, cashAmount, upiAmount, udhaarAmount, status, createdAt, items]
      properties:
        id:
          type: integer
        billNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        cashAmount:
          type: number
        upiAmount:
          type: number
        udhaarAmount:
          type: number
        discountAmount:
          type: number
        status:
          type: string
          enum: [active, cancelled]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillLineItem"
    BillLineItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
        discountAmount:
          type: number
    BillInput:
      type: object
      required: [items, totalAmount, cashAmount, upiAmount, udhaarAmount]
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillItemInput"
        totalAmount:
          type: number
        cashAmount:
          type: number
        upiAmount:
          type: number
        udhaarAmount:
          type: number
        discountAmount:
          type: number
        notes:
          type: string
    BillItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number
        discountAmount:
          type: number
    CancelBillInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── CUSTOMER ─────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, udhaarBalance, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        udhaarBalance:
          type: number
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
    CustomerDetail:
      type: object
      required: [id, name, udhaarBalance, status, createdAt, ledger]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        udhaarBalance:
          type: number
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
        ledger:
          type: array
          items:
            $ref: "#/components/schemas/UdhaarLedgerEntry"
    UdhaarLedgerEntry:
      type: object
      required: [id, customerId, entryType, amount, balanceAfter, description, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        entryType:
          type: string
          enum: [debit, credit]
        amount:
          type: number
        balanceAfter:
          type: number
        description:
          type: string
        billId:
          type: ["integer", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        openingBalance:
          type: number
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    UdhaarPaymentInput:
      type: object
      required: [amount, paymentMode]
      properties:
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi]
        notes:
          type: string

    # ─── SUPPLIER ─────────────────────────────────────────────────────────────
...